research-bot/requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
anthropic>=0.40
notion-client>=2.2
pydantic>=2.0
pyyaml>=6.0
apscheduler>=3.10
python-dotenv>=1.0
click>=8.1
lxml>=5.0
fastapi>=0.115
uvicorn[standard]>=0.32
jinja2>=3.1
